H3: isolate throughput bench and helpers
