+++ quinn-h3/Cargo.toml
@@ -61,5 +61,5 @@ name = "h3_server"
 name = "h3_client"
 
 [[bench]]
-name = "bench"
+name = "throughput"
 harness = false
